+++ mlxtend/classifier/tests/test_stacking_classifier.py
@@ -10,6 +10,7 @@ from mlxtend.classifier import StackingClassifier
 from mlxtend.externals.estimator_checks import NotFittedError
 from scipy import sparse
 from sklearn.linear_model import LogisticRegression
+from sklearn.linear_model import PassiveAggressiveClassifier
 from sklearn.naive_bayes import GaussianNB
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.neighbors import KNeighborsClassifier
@@ -515,3 +516,25 @@ def test_clone():
                                meta_classifier=lr,
                                store_train_meta_features=True)
     clone(stclf)
+
+
+def test_decision_function():
+    np.random.seed(123)
+
+    # PassiveAggressiveClassifier has no predict_proba 
+    meta = PassiveAggressiveClassifier(random_state=42)
+    clf1 = RandomForestClassifier(n_estimators=10)
+    clf2 = GaussianNB()
+    sclf = StackingClassifier(classifiers=[clf1, clf2],
+                              use_probas=True,
+                              meta_classifier=meta)
+
+    # binarize target
+    y2 = y > 1
+    scores = cross_val_score(sclf,
+                             X,
+                             y2,
+                             cv=5,
+                             scoring='roc_auc')
+    scores_mean = (round(scores.mean(), 2))
+    assert scores_mean == 0.96, scores_mean